pyproject.toml
[project]
name = "tracknetv4-padel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scipy>=1.16.3",
    "torch~=2.9.1",
    "torchvision~=0.24.1",
    "tqdm~=4.67.1",
    "wandb~=0.23.1",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
